IRA/requirements.txt
streamlit
groq
python-dotenv
tqdm
PyPDF2
python-docx
pandas